+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Add base functionality to apply fixes from magik-lint.
 - Add UnsafeEvaluateInvocationCheck to test for unsafe `unsafe_evaluate()` method calls.
 - Add SwChar16VectorEvaluateInvocationCheck to test for `sw:char16_vector.evaluate()` method calls.
+- Fix setting `magik.typing.index*Usages` internally in `magik-language-server`.
 
 0.9.1 (2024-03-13)